Calculator

Description

This is an implementation of a simple calculator with operations: addition, subtraction, multiplication, division and memory.

Data Types

data Incr t

Constructors

Incr Int t

data Recall t

Constructors

Recall (Int -> t)

class Functor f => Run f where

type class for execution

Methods

runAlgebra :: f (Mem -> (a, Mem)) -> Mem -> (a, Mem)

handles appropriate functions

Executions

run :: Run f => Term f a -> Mem -> (a, Mem)

Run. This makes our calculator executable for incr and recall.

Functions

incr :: Incr :<: f => Int -> Term f ()

Incrementing the number

recall :: Recall :<: f => Term f Int

Restoring (Recalling) the number stored in memory

tick :: Term (Recall :+: Incr) Int

Tick. Reads the number from memory (with recall) and increments it (with incr).