Expr

Description

This is an implementation of expressions described in a scientific paper titled Data types `a la carte.

Data types

data Expr f

Constructors

In (f (Expr f))

data Val a

Constructors

Val Int

data Add e

Constructors

Add e e

data Mul x

Instance for functor used for multiplication

Constructors

Mul x x

data Div x

Instance for functor used for division (int divison)

Constructors

Div x x

data Mod x

Instance for functor used for modulo

Constructors

Mod x x

data (f :+: g) e infixr 7

Coproduct of two signatures

Constructors

Inl (f e)

Inr (g e)

class (Functor sub, Functor sup) => sub :<: sup where

With this type class, we won't have to write injections using Inl and Inr

Methods

inj :: sub a -> sup a

injection

Classes

class Render f where

Rendering the output

Methods

render :: Render g => f (Expr g) -> String

class Functor f => Eval f where

Class for Evaluating Expr

Methods

evalAlgebra :: f Int -> Int

Appropriate evaluating of given Integer

Basic operators

val :: Val :<: f => Int -> Expr f

Set Expr from Int

(.+) :: Add :<: f => Expr f -> Expr f -> Expr f infixl 6

Sum two Expr

various functions

eval :: Eval f => Expr f -> Int

Evaluation of expression

compose :: [a -> a] -> a -> a

pretty :: Render f => Expr f -> String

Pretty the output

foldExpr :: Functor f => (f a -> a) -> Expr f -> a

Folds the expression

various examples

addExample :: Expr (Val :+: Add)

Basic example, of how awful would it be, if we weren't come up with more clever solution

rezSestevanje :: Expr ((:+:) Add Val)

Example of addition

rezMnozenje :: Expr ((:+:) Val Mul)

Example of multiplication

rezDeljenje :: Expr ((:+:) Val Div)

Example of division

rezModul :: Expr ((:+:) Val Mod)

Example of modulo

rezSestevanjeMnozenje :: Expr ((:+:) Val ((:+:) Add Mul))

Example of multiplication and addition