Main

Description

This is an implementation of a scientific paper titled Data types `a la carte. It includes a simple calculator with operations: addition, subtraction, multiplication, division and memory It also includes two types of IO operations. Teletype operations - reading characters from keyboard and printing them on monitor and FileSystem operations - reading and writing from external file