ReaderWriter

Description

This is an implementation of two types of IO operations. Teletype operations - reading characters from keyboard and printing them on monitor and FileSystem operations - reading and writing from external file

Data Types

data Teletype a

Data Teletype

Constructors

GetChar (Char -> a)

PutChar Char a

data FileSystem a

Data FileSystem

Constructors

ReadFile FilePath (String -> a)

WriteFile FilePath String a

class Functor f => Exec f where

Methods

execAlgebra :: f (IO a) -> IO a

Execution Algebra, that executes the appropriate IO action

Executions

exec :: Exec f => Term f a -> IO a

Executions

Functions

cat :: FilePath -> Term (Teletype :+: FileSystem) ()

readFile executable

wr :: FilePath -> String -> Term (Teletype :+: FileSystem) ()

zapisiFile executable | wr potDoDatotek+imeDatoteke VsebinaDatoteke

getChar :: Teletype :<: f => Term f Char

getChar (read char from keyboard)

putChar :: Teletype :<: f => Char -> Term f ()

putChar (display char on screen)

readFile :: FileSystem :<: f => FilePath -> Term f String

readFile (read the content from the external file)

writeFile :: FileSystem :<: f => FilePath -> String -> Term f ()

writeFile (write string to external file)

getLine :: Teletype :<: Teletype => Term Teletype String

getLine executable (wait and read user's input until enter is pressed)

putLine :: Teletype :<: Teletype => String -> Term Teletype ()

putLine executable (display string to terminal)

dajChar :: Teletype :<: Teletype => Char -> Term Teletype ()

putChar executable (display char to terminal)

dobiChar :: Teletype :<: Teletype => Term Teletype Char

getChar executable (wait and read user's input (one char only))