Term

Description

This is an implementation of two types of IO operations. Teletype operations - reading characters from keyboard and printing them on monitor and FileSystem operations - reading and writing from external file

Documentation

data Term f a

Constructors

Pure a

Impure (f (Term f a))

foldTerm :: Functor f => (a -> b) -> (f b -> b) -> Term f a -> b

inject2 :: g :<: f => g (Term f a) -> Term f a